refactoring api files
