+++ Dockerfile
@@ -14,4 +14,4 @@ COPY . .
 EXPOSE 8080
 
 # Command to run the FastAPI application
-CMD ["uvicorn", "challenge.api:app", "--host", "0.0.0.0", "--port", "8080"]
+CMD ["uvicorn", "challenge.api.api:app", "--host", "0.0.0.0", "--port", "8080"]
