+++ challenge/__init__.py
@@ -1,3 +1,1 @@
-from challenge.api import app
 
-application = app
