+++ challenge/api/__init__.py
@@ -1,0 +1,3 @@
+from challenge.api.api import app
+
+application = app
